Adding an API for Titansgrave
This API will automatically total all rolled die and modifiers then return it to the chat box. It will also return the number of stunt points the user rolled on the previous roll. If this API is used with the new Titansgrave character it will also return if the user hit the target number that a spell requires to be casted.