Macro bug
Hi!

I found a bug for the macro functionality. When a stat roll is done through a macro the "Beaten diff." ends up being weird. Out by a factor of 10 on difficulties of 0. However if the character sheet api roll button is used this error doesn't happen.

After looking through things I found that the skillLevel var is never parsed as an int inside of the statRoll function. Added a parseInt before it's used and the error no longer happens from the macro.

Cheers,

Rene