Fixed death saving throws being triggered.
With the charactermancer update, the global save modifier field has been added to the death saving throw template by default, even when no global save modifier is used ( {{global=}} ). This is causing the companion script to think that the is a global modifier roll, but errors because no roll is found.  This update will ensure that the handledeathsave method only triggers when there actually is a roll in the global save field.