W&G 3.1 Revised Edition Update (#974)
* W&G Die Roller Update v3.0: WrathAndGlory_DieRoller.js

Added Ability to determine if a test generates Glory, Ruin, Complication, or Critical based on what is rolled on the Wrath die and the role (GM or Player) of the person making the roll.

* W&G Die Roller Update v3.0: Readme.md

Updated the following in the Readme.md
- version number
- changelog

* W&G Die Roller Update v3.0: script.json

Updated the following in the script.json
- version number
- slight tweak to description text

* W&G Die Roller Update - WrathAndGlory_DieRoller.js + folder 3.1

Updated die roller to handle concatenation of multiple wrath die results; revised edition will have multiple wrath die under specific conditions.  The die roller reports on when a result of a 1 or 6 is rolled.  Replaced text with images and updated help.

* W&G 3.1 Revised Edition Update

Updated release notes.

* W&G 3.1 Revised Edition Update

Added version 3.1 - update for revised edition.