Fumbler: expanded robustness and updated README
Increased robustness of text parsing allowing fumble to work even with weird formatting (if html tags somehow got into the handout) and allowed user to use tab characters in handout.

Also bounded the dice roll so it shouldn't give a value that is greater than the maximum value in the fumble chart.