Keep teleport from activating if a token has not left the teleport pad.
This change checks the "lastmove" and subjects it to the same check as the current token position on drop. If the "lastmove" is within the bounds of the same teleport pad to-be-activated, the teleport instead stops, as the common circumstance is a token teleported to a large teleport pad, and not yet out of the bounds of it.