Update cypher_systems_by_roll20.js
Add XP as a valid stat to be modified by the "!cypher-modstat" function. The statcost is added (or subtracted if negative) from the character's XP total. This is useful in conjunction with the use of XP tokens as a deck, especially in conjunction with Keith Curtis's "Dealer" script (which makes dealing cards to and taking cards from players into API calls) to keep the user's number of XP card tokens and the user's XP total in sync.