Submitting 2 related scripts for 1-click
PLEASE INCLUDE BOTH APILOGIC AND LIBINLINE IN THE 1-CLICK

APILogic is a utility script that gives the user chat-level logical structures (if, elseif, else) to conditionally evaluate the command line to include/exclude  parts based on conditional evaluations.  It REQUIRES libInline and expects it to be named "libinline" in the approved.yaml.

libInline is a utility script that provides an interface for easy interaction with the inlinerolls array of the message object, including extracting dice sets and/or producing a well-formed roll-tip.