Initial load
v1.014 If players who control characters are not online, send messages for that character to the GM.
v1.015 Fixed a problem with Haste which caused subsequent attack initiatives to fail. (also fixed Haste-end effect)
v1.016 Changed spell initiative menu to display the correct number of spells at every level up to 7 for priests or 9 for wizards
v1.017 Corrected inconsistencies in the name of init_submitVal, which caused initiative carry-over to not work.
v1.018 Updated to deal with new spell usage, retaining name but disabling selection button for cast-value of 0
v1.019 Fixed error with doing initiative for the second MI in any character's MI Bag, which would always cause an API crash.
v1.020 Amend attrLookup() and setAttr() to use fields object more efficiently