New Effects, relative aoe measurements, fix errors (#1569)
* New Effects, relative aoe measurements, fix errors

Added effects to turn on and off Underwater Infravision.  Added error messages when editing or moving statuses and no tokens are selected.  Sent --redo message for each to affected token to InitMaster API when clearing the turn order.  Allow relative measurements to caster token for Area of Effects (using +/-).  Fix errors in getTokenValues().

* Corrected script.json