Create Dimension Bag
This script creates a bag, sack, or hole handout that will track Items, amount, weight and overall weight.
To add items to the bag, sack or hole you will need to use one of these macros:
!dimensionbag add ["Bag Name"] [Amount] ["Item"] [Weight]
!sack add ["Sack Name"] [Amount] ["Item"] [Weight]
!hole add ["Hole Name"] [Amount] ["Item"] [Weight]
To remove items from the bag, sack or hole use these macros:
!dimensionbag remove ["Bag Name"] [Amount] ["Item"]
!sack remove ["Sack Name"] [Amount] ["Item"]
!hole remove ["Hole Name"] [Amount] ["Item"]
If the weight threshold is reached a message is posted in chat to alert the players that the bag or sack is overloaded and its contents are lost. The hole has unlimited weight but its up to the GM to determine if too much space is taken up to be able to fit anymore in.