Update Teleport to 1.2.4
This corrects a flaw in recording entries inside of tokens from JavaScript Objects to JSON data and associated methods to encode and parse this JSON data recorded on tokens. It also introduces token creation for uploaded and Marketplace tokens, and duplicates states from a teleporting token to an existing token on another page.